backend/pyproject.toml
[project]
name = "oneul-naliga-backend"
version = "0.1.0"
description = "주식 차트 분석 백엔드 API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.6",
    "python-dotenv>=1.0.1",
    "supabase>=2.7.4",
    "httpx>=0.27.2",
    "python-telegram-bot>=21.5",
    "pydantic-settings>=2.4.0",
    "websockets>=15.0.1",
    "pytz>=2026.1.post1",
    "scipy>=1.17.1",
    "numpy>=2.4.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
